requirements.txt
Package              Version
-------------------- ------------
annotated-types      0.5.0
anyio                3.7.1
argon2-cffi          23.1.0
argon2-cffi-bindings 21.2.0
asgiref              3.7.2
attrs                23.1.0
backcall             0.2.0
beautifulsoup4       4.12.2
bleach               6.0.0
blis                 0.7.11
catalogue            2.0.10
certifi              2023.7.22
cffi                 1.15.1
charset-normalizer   3.3.2
ChatterBot           1.0.8
click                8.1.7
cloudpathlib         0.16.0
comm                 0.1.4
confection           0.1.3
cycler               0.11.0
cymem                2.0.8
debugpy              1.7.0
decorator            5.1.1
defusedxml           0.7.1
distro               1.8.0
Django               3.2.23
djangorestframework  3.14.0
entrypoints          0.4
es-core-news-sm      3.1.0
exceptiongroup       1.1.3
fastjsonschema       2.18.1
fonttools            4.38.0
h11                  0.14.0
httpcore             0.17.3
httpx                0.24.1
idna                 3.4
importlib-metadata   6.7.0
importlib-resources  5.12.0
ipykernel            6.16.2
ipython              7.34.0
ipython-genutils     0.2.0
ipywidgets           8.1.1
jedi                 0.19.1
Jinja2               3.1.2
joblib               1.3.2
jsonschema           4.17.3
jupyter              1.0.0
jupyter_client       7.4.9
jupyter-console      6.6.3
jupyter_core         4.12.0
jupyter-server       1.24.0
jupyterlab-pygments  0.2.2
jupyterlab-widgets   3.0.9
kiwisolver           1.4.5
langcodes            3.3.0
llvmlite             0.39.1
MarkupSafe           2.1.3
mathparse            0.1.2
matplotlib           3.5.3
matplotlib-inline    0.1.6
mistune              3.0.2
murmurhash           1.0.10
nbclassic            1.0.0
nbclient             0.7.4
nbconvert            7.6.0
nbformat             5.8.0
nest-asyncio         1.5.8
notebook             6.5.6
notebook_shim        0.2.3
numba                0.56.4
numpy                1.21.6
openai               1.3.4
packaging            23.2
pandas               1.3.5
pandocfilters        1.5.0
parso                0.8.3
pathy                0.10.3
patsy                0.5.3
pexpect              4.8.0
pickleshare          0.7.5
Pillow               9.5.0
pip                  23.3.1
pkgutil_resolve_name 1.3.10
plotly               5.18.0
plotly-express       0.4.1
preshed              3.0.9
prometheus-client    0.17.1
prompt-toolkit       3.0.39
psutil               5.9.6
ptyprocess           0.7.0
pybliometrics        3.5.2
pycparser            2.21
pydantic             1.10.13
pydantic_core        2.14.3
Pygments             2.16.1
pynndescent          0.5.11
pyparsing            3.1.1
PyPDF2               3.0.1
pyrsistent           0.19.3
python-dateutil      2.8.2
pytz                 2023.3.post1
pyzmq                24.0.1
qtconsole            5.4.4
QtPy                 2.4.0
requests             2.31.0
scikit-learn         1.0.2
scipy                1.7.3
Send2Trash           1.8.2
setuptools           62.6.0
six                  1.16.0
smart-open           6.4.0
sniffio              1.3.0
soupsieve            2.4.1
spacy                3.7.2
spacy-legacy         3.0.12
spacy-loggers        1.0.5
SQLAlchemy           1.3.24
sqlparse             0.4.4
srsly                2.4.8
statsmodels          0.13.5
tenacity             8.2.3
terminado            0.17.1
thinc                8.2.1
threadpoolctl        3.1.0
tinycss2             1.2.1
tornado              6.2
tqdm                 4.66.1
traitlets            5.9.0
typer                0.4.2
typing_extensions    4.7.1
umap-learn           0.5.5
urllib3              2.0.7
wasabi               0.10.1
wcwidth              0.2.8
weasel               0.3.4
webencodings         0.5.1
websocket-client     1.6.1
wheel                0.37.1
widgetsnbextension   4.0.9
zipp                 3.15.0